+++ lists/safestack.go
@@ -7,26 +7,22 @@ import (
 
 // A SafeStack is an abstract data type that serves as a collection of elements with two main operations:
 //
-// # SafeStack is threadsafe
-//
-// # Please note that while the type itself is safe the data should handle its own thread safety
-//
 // Push, which adds an element to the collection, and
 // Pop, which removes the most recently added element.
 // Additionally, a peek operation can, without modifying the stack, return the value of the last element added.
 type SafeStack[T any] struct {
-	mu          sync.RWMutex
 	curBuffSize uint
-	head        *snode[T]
-	tail        *snode[T]
+	index       uint
+	head        *arrnode[T]
+	mu          sync.RWMutex
 }
 
 // Constructs a new Stack with elements of type T
 func NewSafeStack[T any]() *SafeStack[T] {
 	return &SafeStack[T]{
 		curBuffSize: 0,
-		head:        nil,
-		tail:        nil,
+		head:        newArrayNode[T](nil),
+		index:       999,
 	}
 }
 
@@ -35,15 +31,17 @@ func (r *SafeStack[T]) Push(element T) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 
-	if r.curBuffSize == 0 {
-		r.head = newSingleNode[T](element, nil)
-		r.tail = r.head
-		r.curBuffSize++
-	} else {
-		r.curBuffSize++
-		newItem := newSingleNode(element, r.head)
-		r.head = newItem
+	if r.curBuffSize > 0 {
+		if r.index == 0 {
+			r.index = 999
+			newNode := newArrayNode[T](r.head)
+			r.head = newNode
+		} else {
+			r.index--
+		}
 	}
+	r.curBuffSize++
+	r.head.write(element, int(r.index))
 }
 
 // Removes the most recently added element T from the stack and returns it. Complexity is O(1)
@@ -51,19 +49,23 @@ func (r *SafeStack[T]) Pop() (T, error) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 
+	var result T
 	if r.curBuffSize == 0 {
-		var result T
 		return result, errors.New("empty list")
 	}
 
-	ret := r.head.data
-	r.head = r.head.next
 	r.curBuffSize--
-	if r.curBuffSize == 0 {
-		r.head = nil
-		r.tail = nil
+	result = r.head.read(int(r.index))
+	r.index++
+
+	if r.index > 999 {
+		r.index = 0
+		if r.head.next != nil {
+			r.head = r.head.next
+		}
 	}
-	return ret, nil
+
+	return result, nil
 }
 
 // Checks to see if the stack is empty.
@@ -72,19 +74,18 @@ func (r *SafeStack[T]) Pop() (T, error) {
 func (r *SafeStack[T]) IsEmpty() bool {
 	r.mu.RLock()
 	defer r.mu.RUnlock()
+
 	return r.curBuffSize == 0
 }
 
 // The Peek operation returns, without modifying the stack, the value of the last element T added
 func (r *SafeStack[T]) Peek() (T, error) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
+	var result T
 	if r.curBuffSize == 0 {
-		var result T
 		return result, errors.New("empty list")
 	}
 
-	return r.head.data, nil
+	return r.head.read(int(r.index)), nil
 }
 
 // Return a slice representation of the current state of the stack
@@ -92,16 +93,31 @@ func (r *SafeStack[T]) ToSlice() []T {
 	r.mu.RLock()
 	defer r.mu.RUnlock()
 
+	s := make([]T, 0, r.curBuffSize)
+
 	if r.curBuffSize == 0 {
-		return make([]T, 0)
+		return s
 	}
 
-	s := make([]T, 0, r.curBuffSize)
 	tmp := r.head
+	index := r.index
+	finished := false
+	cnt := 0
 
-	for tmp != nil {
-		s = append(s, tmp.data)
-		tmp = tmp.next
+	for tmp != nil && !finished {
+		s = append(s, tmp.read(int(index)))
+		cnt++
+
+		if index == 999 {
+			index = 0
+			tmp = tmp.next
+		} else {
+			index++
+		}
+
+		if cnt > int(r.curBuffSize) {
+			finished = true
+		}
 	}
 	return s
 }